SearchAPI/requirements_py38.txt
asgiref==3.7.2
astropy==4.2
astroquery==0.4.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.39.4
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.13.1
kiwisolver==1.4.4
matplotlib==3.7.1
mimeparse==0.1.3
more-itertools==9.1.0
numpy==1.21.1
packaging==23.1
Pillow==9.5.0
pycparser==2.21
pyerfa==2.0.0.3
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyvo==1.1
requests==2.31.0
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.6.3
urllib3==2.0.3
webencodings==0.5.1
zipp==3.15.0
